Move back the logic to "suspendmanager"
