+++ suspend.lua
@@ -3,10 +3,10 @@
 -- It can either suspend all jobs, or just jobs that risk blocking others.
 
 local argparse = require('argparse')
-local suspendmanagerUtils = reqscript('internal/suspendmanager/suspendmanager-utils')
+local suspendmanager = reqscript('suspendmanager')
 
 local help, onlyblocking = false, false
-argparse.processArgsGetopt(args, {
+argparse.processArgsGetopt({...}, {
     {'h', 'help', handler=function() help = true end},
     {'b', 'onlyblocking', handler=function() onlyblocking = true end},
 })
@@ -16,8 +16,8 @@ if help then
     return
 end
 
-suspendmanagerUtils.foreach_construction_job(function (job)
-    if not onlyblocking or suspendmanagerUtils.isBlocking(job) then
-        suspendmanagerUtils.suspend(job)
+suspendmanager.foreach_construction_job(function (job)
+    if not onlyblocking or suspendmanager.isBlocking(job) then
+        suspendmanager.suspend(job)
     end
 end)
